Integrate v2 types
Renames User struct to UserInfo.
