+++ src/account/auth/test_notifs.rs
@@ -1,5 +1,5 @@
 use crate::{
-    constants::merkle::API_ROOT, types::account::auth::secret::Secret, types::v2::user::UserRoot,
+    constants::merkle::API_ROOT, types::account::auth::secret::Secret, types::user::user::UserRoot,
     Farcaster,
 };
