+++ src/types/user/mod.rs
@@ -2,3 +2,5 @@ pub mod followers;
 pub mod notifications;
 pub mod user;
 pub mod verified_address;
+
+pub use user::UserInfo;
