+++ src/types/user/user.rs
@@ -1,59 +1,56 @@
 use serde::{Deserialize, Serialize};
-use serde_json::Value;
 
-/// # User Root Struct
-///
-/// When getting a user, we need structs to Deserialize the JSON into
-///
-/// This struct holds one value
-///
-/// * `result: Result`
-///     - Result holds a User type, which itself holds a lot of the user data
-///     - To view more about the data types, explore the many structs in this file.
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Root {
+#[derive(Debug, Deserialize, Serialize)]
+pub struct UserRoot {
     pub result: Result,
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Result {
-    pub user: User,
+    pub user: UserInfo,
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct User {
-    pub address: String,
+#[derive(Debug, Deserialize, Serialize)]
+pub struct UserInfo {
+    pub fid: i64,
     pub username: String,
+    #[serde(rename = "displayName")]
     pub display_name: String,
     pub avatar: Avatar,
+    #[serde(rename = "followerCount")]
     pub follower_count: i64,
+    #[serde(rename = "followingCount")]
     pub following_count: i64,
+    #[serde(rename = "isViewerFollowing")]
     pub is_viewer_following: bool,
+    #[serde(rename = "isFollowingViewer")]
     pub is_following_viewer: bool,
     pub profile: Profile,
-    pub referrer_username: Option<String>,
-    pub viewer_can_send_direct_casts: bool,
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Avatar {
     pub url: String,
+    #[serde(rename = "isVerified")]
     pub is_verified: bool,
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Profile {
     pub bio: Bio,
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Bio {
     pub text: String,
-    pub mentions: Vec<Value>,
+    pub mentions: Vec<Mention>,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct Mention {
+    pub avatar: Avatar,
+    pub fid: i64,
+    pub username: String,
+    #[serde(rename = "displayName")]
+    pub display_name: String,
 }
