+++ src/users/mod.rs
@@ -1,5 +1,5 @@
 //pub mod get_casts;
 //pub mod get_followers;
 //pub mod get_notifications;
-//pub mod get_user;
+pub mod get_user;
 //pub mod get_verified_address;
